Refactor: 레파지토리 & discount 인터페이스 관련 네이밍을 수정하라
