+++ src/main/java/org/prgrms/kdt/app/command/CustomerVoucherCreateCommand.java
@@ -1,6 +1,5 @@
 package org.prgrms.kdt.app.command;
 
-import java.time.LocalDateTime;
 import java.util.UUID;
 import org.prgrms.kdt.app.io.Console;
 import org.prgrms.kdt.model.wallet.WalletVO;
