+++ src/main/java/org/prgrms/kdt/model/discount/DiscountPolicy.java
@@ -1,6 +1,6 @@
 package org.prgrms.kdt.model.discount;
 
-public interface DiscountStrategy {
+public interface DiscountPolicy {
     long discount(long beforeDiscount, long discount);
 
     void validateBeforeDiscountAmount(long beforeDiscount);
