+++ src/main/java/org/prgrms/kdt/model/discount/PercentDiscountPolicy.java
@@ -3,12 +3,12 @@ package org.prgrms.kdt.model.discount;
 import java.text.MessageFormat;
 import org.prgrms.kdt.exception.InvalidDataException;
 
-public class FixedDiscountStrategy implements DiscountStrategy {
+public class PercentDiscountPolicy implements DiscountPolicy {
 
     @Override
     public long discount(long beforeDiscount, long discount) {
-        validateDiscountAmount(beforeDiscount);
-        return beforeDiscount > discount ? beforeDiscount - discount : 0;
+        validateBeforeDiscountAmount(beforeDiscount);
+        return (beforeDiscount * (100 - discount)) / 100;
     }
 
     @Override
@@ -22,13 +22,9 @@ public class FixedDiscountStrategy implements DiscountStrategy {
 
     @Override
     public void validateDiscountAmount(long discount) {
-        if (discount < 0) {
+        if (discount < 0 || discount > 100) {
             throw new InvalidDataException(
-                MessageFormat.format("fixed discount amount should be positive: {0}", discount));
-        }
-        if (discount == 0) {
-            throw new InvalidDataException(
-                MessageFormat.format("fixed discount amount should not be zero: {0}", discount));
+                MessageFormat.format("percentage discount amount should be 0~100: {0}", discount));
         }
     }
 }
