+++ src/main/java/org/prgrms/kdt/model/voucher/VoucherType.java
@@ -1,19 +1,19 @@
 package org.prgrms.kdt.model.voucher;
 
-import org.prgrms.kdt.model.discount.DiscountStrategy;
-import org.prgrms.kdt.model.discount.FixedDiscountStrategy;
-import org.prgrms.kdt.model.discount.PercentDiscountStrategy;
+import org.prgrms.kdt.model.discount.DiscountPolicy;
+import org.prgrms.kdt.model.discount.FixedDiscountPolicy;
+import org.prgrms.kdt.model.discount.PercentDiscountPolicy;
 
 public enum VoucherType {
-    FIX("1", new FixedDiscountStrategy()),
-    PERCENT("2", new PercentDiscountStrategy());
+    FIX("1", new FixedDiscountPolicy()),
+    PERCENT("2", new PercentDiscountPolicy());
 
     private final String num;
-    private final DiscountStrategy discountStrategy;
+    private final DiscountPolicy discountPolicy;
 
-    VoucherType(String value, DiscountStrategy discountStrategy) {
+    VoucherType(String value, DiscountPolicy discountPolicy) {
         this.num = value;
-        this.discountStrategy = discountStrategy;
+        this.discountPolicy = discountPolicy;
     }
 
 
@@ -21,8 +21,8 @@ public enum VoucherType {
         return num;
     }
 
-    public DiscountStrategy getDiscountStrategy() {
-        return discountStrategy;
+    public DiscountPolicy getDiscountPolicy() {
+        return discountPolicy;
     }
 
     @Override
